% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gens.R
\name{ar_method}
\alias{ar_method}
\title{Acceptance-Rejection method}
\usage{
ar_method(n, f, g, rg, c)
}
\arguments{
\item{n}{number of random numbers}

\item{f}{interest density function}

\item{g}{sampling density function}

\item{rg}{random number generator function with distribution g}

\item{c}{constant or a vector containing the end-points of the interval to be searched
for the maximum constant such that \eqn{f(y)/g(y)<=c} for all y.}
}
\description{
Acceptance-Rejection method
}
