% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gens.R
\name{comp_method}
\alias{comp_method}
\title{Composition Method}
\usage{
comp_method(n, qfns, p)
}
\arguments{
\item{n}{number of random numbers}

\item{qfns}{list of the quantile functions in order}

\item{p}{vector of probabilities in order}
}
\description{
Composition Method
}
