% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gens.R
\name{comp_method_c}
\alias{comp_method_c}
\title{Composition method (Continuous)}
\usage{
comp_method_c(n, qfy, qfx)
}
\arguments{
\item{n}{number of random numbers}

\item{qfy}{quantile function of Y}

\item{qfx}{quantile function of X|Y, where the first argument is the probability and the second
is y}
}
\description{
Composition method (Continuous)
}
