% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gens.R
\name{comp_method_d}
\alias{comp_method_d}
\title{Composition method (Discrete)}
\usage{
comp_method_d(n, a, qpx, qqx)
}
\arguments{
\item{n}{number of random numbers}

\item{a}{double constant}

\item{qpx}{quantile function of px}

\item{qqx}{quantile function of qx}
}
\description{
\deqn{Pr(X = x) = a * p_{x} + (1-a) * q_{x}}
}
