% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gens.R
\name{convol_method}
\alias{convol_method}
\title{Convolution method}
\usage{
convol_method(n, y, qfns)
}
\arguments{
\item{n}{number of random numbers}

\item{y}{function of the definition of y in terms of \eqn{x_i}}

\item{qfns}{list of the \eqn{k} quantile functions in order}
}
\description{
\deqn{Y ~ \sum (\beta_i * X_i) , 1 \le i \le k}
}
