% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gens.R
\name{lcg}
\alias{lcg}
\title{Linear Congruential Generator}
\usage{
lcg(n, seed, a, c, m)
}
\arguments{
\item{n}{number of random numbers}

\item{seed}{Seed}

\item{a}{positive integer}

\item{c}{positive integer}

\item{m}{positive integer}
}
\description{
\deqn{x_{n} = (ax_{n-1} + c) modulo m}
}
