% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\docType{data}
\name{models}
\alias{models}
\alias{real_support}
\alias{positive_real_support}
\alias{zero_one_support}
\title{Model support classifications}
\format{
An object of class \code{character} of length 10.

An object of class \code{character} of length 11.

An object of class \code{character} of length 3.
}
\usage{
real_support

positive_real_support

zero_one_support
}
\description{
\code{\link[univariateML]{univariateML_models}} grouped by type of support.
}
\details{
\code{real_support}: Cauchy, Gumbel, Laplace, Logistic, Normal, Student t, Generalized Error,
Skew Normal, Skew Student t, Skew Generalized Error

\code{positive_real_support}: Beta prime, Exponential, Gamma, Inverse gamma, Inverse Gaussian,
Inverse Weibull, Log-logistic, Log-normal, Lomax, Rayleigh, Weibull

\code{zero_one_support}: Beta, Kumaraswamy, Logit-normal

Other models:
\itemize{
\item \code{lgamma}: Log-gamma (1, \eqn{\infty})
\item \code{pareto}: Pareto [b, \eqn{\infty})
\item \code{power}: Power [0,a)
\item \code{unif}: Uniform (min, max)
}
}
\keyword{datasets}
