% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_discrete}
\alias{plot_discrete}
\title{Barplot}
\usage{
plot_discrete(v, fun = NULL, args = NULL, ...)
}
\arguments{
\item{v}{numeric vector}

\item{fun}{function to graph}

\item{args}{list of arguments passed to fun}
}
\description{
Plots a barplot with the possibility of overlaying a discrete function
}
