% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{plot_mc_convergence}
\alias{plot_mc_convergence}
\title{Monte Carlo convergence}
\usage{
plot_mc_convergence(x, reference = NULL, ...)
}
\arguments{
\item{x}{numeric vector}

\item{reference}{true value to be estimated through Monte Carlo simulations}

\item{...}{other arguments passed on to \code{\link[ggplot2]{geom_path}}}
}
\description{
Line plot of the cumulative mean for a vector of Monte Carlo simulations
}
