% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{plot_models}
\alias{plot_models}
\alias{plot_dml}
\alias{plot_pml}
\alias{plot_qqml}
\title{Plot models}
\usage{
plot_dml(v, object, ..., xlim = NULL)

plot_pml(v, object, ..., xlim = NULL)

plot_qqml(v, object, ...)
}
\arguments{
\item{v}{numeric vector}

\item{object}{resulting object from \code{fit_mlm}}

\item{...}{other arguments passed on to \code{\link[ggplot2]{geom_histogram}} for \code{plot_dml},
to \code{\link[ggplot2]{stat_ecdf}} for \code{plot_pml}, to \code{\link[ggplot2]{stat_qq}} for
\code{plot_qqml}}

\item{xlim}{optionally, restrict the range of the function to this range}
}
\description{
Returns a list of plots, each with the data and the fitted model overlaid.
}
\details{
The returned list allows, in conjunction with \code{\link[patchwork]{wrap_plots}} or another
function from any library of your choice, to combine the resulting plots into one, or simply
edit/show each plot separately.
}
