% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gens.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Multivariate Normal Distribution}
\usage{
rmvnorm(n, u, S)
}
\arguments{
\item{n}{number of random numbers}

\item{u}{mean vector}

\item{S}{covariance matrix}
}
\description{
Random number generator for the multivariate normal distribution with mean equal to
\eqn{u} and covariance matrix \eqn{S}.
}
