% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gens.R
\name{rnorm_boxmuller}
\alias{rnorm_boxmuller}
\title{Box–Muller transform}
\usage{
rnorm_boxmuller(n)
}
\arguments{
\item{n}{number of random numbers}
}
\description{
Random number generator for pairs of independent, standard, normal distributions.
}
