% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gens.R
\name{wh}
\alias{wh}
\title{Wichmann-Hill (WH) Random Number Generator}
\usage{
wh(n, seed, a, c, m)
}
\arguments{
\item{n}{number of random numbers}

\item{seed}{3 seeds}

\item{a}{3 positive integers}

\item{c}{3 positive integers}

\item{m}{3 positive integers}
}
\description{
The Wichmann-Hill (WH) algorithm is an extension of linear congruential. WH uses three linear
congruential equations, adds the results, then takes the decimal part of the sum. For example,
if the three WH equations give 0.68, 0.92, and 0.35, then the final result random number is
decimal(0.78 + 0.92 + 0.45) = decimal(2.95) = 0.95.
}
