% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gens.R
\name{comp_method}
\alias{comp_method}
\title{Composition method}
\usage{
comp_method(n, qfns, p)
}
\arguments{
\item{n}{number of random numbers}

\item{qfns}{list of the \eqn{k} quantile functions in order}

\item{p}{vector of \eqn{k} probabilities in order}
}
\description{
\deqn{F(x) = \sum (p_i * F_i(x)) , 1 \le i \le k}
}
