% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_continuous}
\alias{plot_continuous}
\title{Histogram}
\usage{
plot_continuous(v, fun = NULL, args = NULL, fill = "#C3395B", ...)
}
\arguments{
\item{v}{numeric vector}

\item{fun}{function to plot}

\item{args}{list of arguments passed to fun}

\item{...}{other arguments passed on to \code{\link[ggplot2]{geom_histogram}}}
}
\description{
Plots a histogram with the possibility of overlaying a function
}
