% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_discrete}
\alias{plot_discrete}
\title{Barplot}
\usage{
plot_discrete(v, fun = NULL, args = NULL, fill = "#C3395B", n = 5, ...)
}
\arguments{
\item{v}{numeric vector}

\item{fun}{function to graph}

\item{args}{list of arguments passed to fun}

\item{n}{desired number of breaks in x axis. You may get slightly more or fewer breaks
that requested.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{stat_count}}}
}
\description{
Plots a barplot with the possibility of overlaying a discrete function
}
